회사 상세보기, 리뷰 페이지 디테일 작업

+++ src/main/java/com/kh/khblind/company/entity/CompanyReviewDto.java
@@ -16,6 +16,6 @@ public class CompanyReviewDto {
 	private String reviewMerit;
 	private String reviewDemerit;
 	private Date reviewDate;
-	private double reviewRate;
+	private int reviewRate;
 	private int jobCategoryNo;
 }
